change class UrlDto-LongUrlDto
