+++ src/main/java/nadja/url_shortener/service/StringShortenerHelper.java
@@ -1,10 +1,8 @@
 package nadja.url_shortener.service;
 
-import org.springframework.stereotype.Service;
-
 import java.nio.ByteBuffer;
 import java.util.Base64;
-//@Service
+
 public class StringShortenerHelper {
 
     public static String get_shortURL(String longUrl) {
