+++ src/test/java/nadja/url_shortener/controller/RestUrlControllerTest.java
@@ -1,13 +1,12 @@
 package nadja.url_shortener.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import nadja.url_shortener.dto.UrlDto;
+import nadja.url_shortener.dto.LongUrlDto;
 import nadja.url_shortener.entity.Url;
 import nadja.url_shortener.repo.IUrlRepo;
 import nadja.url_shortener.service.ShortenerService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
@@ -22,7 +21,6 @@ import java.time.LocalDate;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
-import static org.springframework.http.converter.json.Jackson2ObjectMapperBuilder.json;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
@@ -32,7 +30,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @AutoConfigureMockMvc
 class RestUrlControllerTest {
     @Autowired
-private ObjectMapper objectMapper;
+    private ObjectMapper objectMapper;
     @Autowired
     private MockMvc mockMvc;
     @Autowired
@@ -48,14 +46,14 @@ private ObjectMapper objectMapper;
         MockitoAnnotations.openMocks(this);
         this.mockMvc = MockMvcBuilders.webAppContextSetup(this.wac).dispatchOptions(true).build();
     }
+
     @Test
     void convertLongUrl() throws Exception {
-        Url urlTest=new Url(1,"http://microsoft.com","Yf74Nb4",0, LocalDate.now().plusDays(3));
-        when(shortenerService.createUrl(any(UrlDto.class))).thenReturn(urlTest);
+        Url urlTest = new Url(1, "http://microsoft.com", "Yf74Nb4", 0, LocalDate.now().plusDays(3));
+        when(shortenerService.createUrl(any(LongUrlDto.class))).thenReturn(urlTest);
 
         mockMvc.perform(post("/shortenerUrl")
-                      //  .accept(MediaType.TEXT_PLAIN )
-        .contentType(MediaType.APPLICATION_JSON)
+                .contentType(MediaType.APPLICATION_JSON)
                 .content("{ \"longUrl\": \"http://microsoft.com\"}"))
                 .andDo(print())
                 .andExpect(status().isOk())
